requirement.txt
certifi==2020.12.5
chardet==4.0.0
fiscalyear==0.3.0
idna==2.10
mysqlclient==2.0.3
patool==1.12
requests==2.25.1
SQLAlchemy==1.3.22
urllib3==1.26.2
xmltodict==0.12.0